Move test suite timings code out of main src
