+++ tests/performance_tests/timings.cc
@@ -1,4 +1,4 @@
-#include <string.h>
+#include <cstring.h>
 #include <errno.h>
 #include <time.h>
 #include <algorithm>
